--- a/tt_metal/common/private_pch.hpp
+++ b/tt_metal/common/private_pch.hpp
@@ -1,0 +1,29 @@
+#pragma once
+
+// API
+#include <tt-metalium/assert.hpp>
+#include <tt-metalium/bfloat16.hpp>
+#include <tt-metalium/device.hpp>
+#include <tt-metalium/host_buffer.hpp>
+#include <tt-metalium/memory_pin.hpp>
+
+// Metalium::Metal::Hardware
+
+// Tracy::TracyClient
+#include "tracy/Tracy.hpp"
+
+// TT::Metalium::HostDevCommon
+#include <hostdevcommon/common_values.hpp>
+
+// yaml-cpp::yaml-cpp
+#include <yaml-cpp/yaml.h>
+
+// Boost::asio
+#include <boost/algorithm/string.hpp>
+#include <boost/asio.hpp>
+#include <boost/asio/post.hpp>
+#include <boost/asio/thread_pool.hpp>
+#include <boost/move/utility_core.hpp>
+
+// tt-logger::tt-logger
+#include <tt-logger/tt-logger.hpp>