--- a/tt_metal/distributed/public_pch.hpp
+++ b/tt_metal/distributed/public_pch.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+// TT::Metalium::Common
+#include <tt_metal/common/public_pch.hpp>
+#include <tt_metal/common/env_lib.hpp>
+#include <tt_metal/common/executor.hpp>
+#include <tt_metal/common/multi_producer_single_consumer_queue.hpp>
+#include <tt_metal/common/thread_pool.hpp>