--- a/tt_metal/impl/CMakeLists.txt
+++ b/tt_metal/impl/CMakeLists.txt
@@ -129,3 +129,14 @@ target_include_directories(
 target_include_directories(impl SYSTEM PRIVATE ${CMAKE_CURRENT_BINARY_DIR}/flatbuffers)
 
 target_compile_options(impl PUBLIC -Wno-int-to-pointer-cast)
+
+if(TT_METAL_BUILD_AGGRESSIVE_PCH)
+    target_precompile_headers(
+        impl
+        PRIVATE
+            ${CMAKE_CURRENT_SOURCE_DIR}/private_pch.hpp
+            ${CMAKE_CURRENT_SOURCE_DIR}/public_pch.hpp
+    )
+else()
+    target_precompile_headers(impl REUSE_FROM TT::CommonPCH)
+endif()