--- a/tt_metal/impl/public_pch.hpp
+++ b/tt_metal/impl/public_pch.hpp
@@ -1,0 +1,42 @@
+#pragma once
+
+// API
+#include <tt-metalium/allocator.hpp>
+#include <tt-metalium/allocator_types.hpp>
+#include <tt-metalium/buffer.hpp>
+#include <tt-metalium/circular_buffer.hpp>
+#include <tt-metalium/circular_buffer_config.hpp>
+#include <tt-metalium/circular_buffer_constants.h>
+#include <tt-metalium/command_queue.hpp>
+#include <tt-metalium/command_queue_interface.hpp>
+#include <tt-metalium/control_plane.hpp>
+#include <tt-metalium/core_coord.hpp>
+#include <tt-metalium/core_descriptor.hpp>
+#include <tt-metalium/data_types.hpp>
+#include <tt-metalium/device.hpp>
+#include <tt-metalium/dispatch_core_common.hpp>
+#include <tt-metalium/distributed_context.hpp>
+#include <tt-metalium/event.hpp>
+#include <tt-metalium/fabric.hpp>
+#include <tt-metalium/hal_types.hpp>
+#include <tt-metalium/kernel.hpp>
+#include <tt-metalium/kernel_types.hpp>
+#include <tt-metalium/lightmetal_binary.hpp>
+#include <tt-metalium/mesh_trace_id.hpp>
+#include <tt-metalium/program.hpp>
+#include <tt-metalium/program_cache.hpp>
+#include <tt-metalium/semaphore.hpp>
+#include <tt-metalium/sub_device.hpp>
+#include <tt-metalium/sub_device_types.hpp>
+#include <tt-metalium/tt_backend_api_types.hpp>
+#include <tt-metalium/vector_aligned.hpp>
+
+// TT::Metalium::Common
+#include <tt_metal/common/public_pch.hpp>
+#include <tt_metal/common/core_assignment.hpp>
+#include <tt_metal/common/multi_producer_single_consumer_queue.hpp>
+
+// Tracy
+#include <common/TracyTTDeviceData.hpp>
+#include <tracy/Tracy.hpp>
+#include <tracy/TracyTTDevice.hpp>