--- a/tt_metal/llrt/private_pch.hpp
+++ b/tt_metal/llrt/private_pch.hpp
@@ -1,0 +1,29 @@
+#pragma once
+
+// API
+#include <tt-metalium/control_plane.hpp>
+
+// HAL::1xx
+
+// Tracy::TracyClient
+#include <tracy/Tracy.hpp>
+
+// nlohmann_json::nlohmann_json
+#include <nlohmann/json.hpp>
+#include <nlohmann/json_fwd.hpp>
+
+// Reflect::Reflect
+#include <reflect>
+
+// yaml-cpp::yaml-cpp
+#include <yaml-cpp/yaml.h>
+
+// enchantum::enchantum
+#include <enchantum/enchantum.hpp>
+#include <enchantum/iostream.hpp>
+
+// TT::Metalium::Common
+#include <tt_metal/common/public_pch.hpp>
+
+// tt-logger::tt-logger
+#include <tt-logger/tt-logger.hpp>