--- a/tt_metal/public_pch.hpp
+++ b/tt_metal/public_pch.hpp
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
 #pragma once
 
 // umd::device