--- a/ttnn/CMakeLists.txt
+++ b/ttnn/CMakeLists.txt
@@ -23,7 +23,11 @@ add_library(ttnn_core ${LIB_TYPE})
 add_library(TTNN::Core ALIAS ttnn_core)
 
 target_compile_definitions(ttnn_core PUBLIC "$<$<CXX_COMPILER_ID:GNU>:DISABLE_NAMESPACE_STATIC_ASSERT>")
-target_precompile_headers(ttnn_core REUSE_FROM TT::CommonPCH)
+if(TT_METAL_BUILD_AGGRESSIVE_PCH)
+    target_precompile_headers(ttnn_core PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/pch.hpp)
+else()
+    target_precompile_headers(ttnn_core REUSE_FROM TT::CommonPCH)
+endif()
 
 set(TENSOR_FLATBUFFER_SCHEMAS
     ${CMAKE_CURRENT_SOURCE_DIR}/core/tensor/flatbuffer/mesh_shape.fbs
@@ -404,7 +408,11 @@ add_library(TTNN::CPP ALIAS ttnncpp)
 add_library(Metalium::TTNNCPP ALIAS ttnncpp) # backwards compatibility. FIXME: remove
 
 target_compile_definitions(ttnncpp PUBLIC "$<$<CXX_COMPILER_ID:GNU>:DISABLE_NAMESPACE_STATIC_ASSERT>")
-target_precompile_headers(ttnncpp REUSE_FROM TT::CommonPCH)
+if(TT_METAL_BUILD_AGGRESSIVE_PCH)
+    target_precompile_headers(ttnncpp REUSE_FROM ttnn_core)
+else()
+    target_precompile_headers(ttnncpp REUSE_FROM TT::CommonPCH)
+endif()
 TT_ENABLE_UNITY_BUILD(ttnncpp)
 
 set_target_properties(
@@ -619,7 +627,18 @@ if(WITH_PYTHON_BINDINGS)
             pybind11::module
             Boost::algorithm
             TT::Metalium
+            FlatBuffers::FlatBuffers
     )
+    if(TT_METAL_BUILD_AGGRESSIVE_PCH)
+        target_precompile_headers(
+            ttnn
+            PRIVATE
+                ${CMAKE_CURRENT_SOURCE_DIR}/pch.hpp
+                ${CMAKE_CURRENT_SOURCE_DIR}/pybind_pch.hpp
+        )
+    else()
+        target_precompile_headers(ttnn REUSE_FROM TT::CommonPCH)
+    endif()
     if(TT_INSTALL)
         install(TARGETS ttnn EXPORT TT-NN LIBRARY COMPONENT ttnn-runtime)
     endif()