--- a/ttnn/cpp/ttnn/operations/core/CMakeLists.txt
+++ b/ttnn/cpp/ttnn/operations/core/CMakeLists.txt
@@ -1,7 +1,11 @@
 add_library(ttnn_op_core ${LIB_TYPE})
 add_library(TTNN::Ops::Core ALIAS ttnn_op_core)
 
-target_precompile_headers(ttnn_op_core REUSE_FROM TT::CommonPCH)
+if(TT_METAL_BUILD_AGGRESSIVE_PCH)
+    target_precompile_headers(ttnn_op_core REUSE_FROM ttnn_core)
+else()
+    target_precompile_headers(ttnn_op_core REUSE_FROM TT::CommonPCH)
+endif()
 TT_ENABLE_UNITY_BUILD(ttnn_op_core)
 
 target_sources(