--- a/ttnn/cpp/ttnn/operations/debug/CMakeLists.txt
+++ b/ttnn/cpp/ttnn/operations/debug/CMakeLists.txt
@@ -1,7 +1,11 @@
 add_library(ttnn_op_debug ${LIB_TYPE})
 add_library(TTNN::Ops::Debug ALIAS ttnn_op_debug)
 
-target_precompile_headers(ttnn_op_debug REUSE_FROM TT::CommonPCH)
+if(TT_METAL_BUILD_AGGRESSIVE_PCH)
+    target_precompile_headers(ttnn_op_debug REUSE_FROM ttnn_core)
+else()
+    target_precompile_headers(ttnn_op_debug REUSE_FROM TT::CommonPCH)
+endif()
 TT_ENABLE_UNITY_BUILD(ttnn_op_debug)
 
 file(GLOB_RECURSE kernels kernels/*)