--- a/ttnn/cpp/ttnn/operations/eltwise/binary/CMakeLists.txt
+++ b/ttnn/cpp/ttnn/operations/eltwise/binary/CMakeLists.txt
@@ -1,7 +1,11 @@
 add_library(ttnn_op_eltwise_binary ${LIB_TYPE})
 add_library(TTNN::Ops::Eltwise::Binary ALIAS ttnn_op_eltwise_binary)
 
-target_precompile_headers(ttnn_op_eltwise_binary REUSE_FROM TT::CommonPCH)
+if(TT_METAL_BUILD_AGGRESSIVE_PCH)
+    target_precompile_headers(ttnn_op_eltwise_binary REUSE_FROM ttnn_core)
+else()
+    target_precompile_headers(ttnn_op_eltwise_binary REUSE_FROM TT::CommonPCH)
+endif()
 TT_ENABLE_UNITY_BUILD(ttnn_op_eltwise_binary)
 set_target_properties(
     ttnn_op_eltwise_binary