--- a/ttnn/cpp/ttnn/operations/eltwise/binary_backward/CMakeLists.txt
+++ b/ttnn/cpp/ttnn/operations/eltwise/binary_backward/CMakeLists.txt
@@ -1,7 +1,11 @@
 add_library(ttnn_op_eltwise_binary_backward ${LIB_TYPE})
 add_library(TTNN::Ops::Eltwise::Binary::Backward ALIAS ttnn_op_eltwise_binary_backward)
 
-target_precompile_headers(ttnn_op_eltwise_binary_backward REUSE_FROM TT::CommonPCH)
+if(TT_METAL_BUILD_AGGRESSIVE_PCH)
+    target_precompile_headers(ttnn_op_eltwise_binary_backward REUSE_FROM ttnn_core)
+else()
+    target_precompile_headers(ttnn_op_eltwise_binary_backward REUSE_FROM TT::CommonPCH)
+endif()
 TT_ENABLE_UNITY_BUILD(ttnn_op_eltwise_binary_backward)
 
 target_sources(ttnn_op_eltwise_binary_backward PRIVATE binary_backward.cpp)