--- a/ttnn/cpp/ttnn/operations/eltwise/binary_ng/CMakeLists.txt
+++ b/ttnn/cpp/ttnn/operations/eltwise/binary_ng/CMakeLists.txt
@@ -1,7 +1,11 @@
 add_library(ttnn_op_eltwise_binary_ng ${LIB_TYPE})
 add_library(TTNN::Ops::Eltwise::Binary::NG ALIAS ttnn_op_eltwise_binary_ng)
 
-target_precompile_headers(ttnn_op_eltwise_binary_ng REUSE_FROM TT::CommonPCH)
+if(TT_METAL_BUILD_AGGRESSIVE_PCH)
+    target_precompile_headers(ttnn_op_eltwise_binary_ng REUSE_FROM ttnn_core)
+else()
+    target_precompile_headers(ttnn_op_eltwise_binary_ng REUSE_FROM TT::CommonPCH)
+endif()
 TT_ENABLE_UNITY_BUILD(ttnn_op_eltwise_binary_ng)
 set_target_properties(
     ttnn_op_eltwise_binary_ng