--- a/ttnn/cpp/ttnn/operations/eltwise/complex/CMakeLists.txt
+++ b/ttnn/cpp/ttnn/operations/eltwise/complex/CMakeLists.txt
@@ -1,7 +1,11 @@
 add_library(ttnn_op_eltwise_complex ${LIB_TYPE})
 add_library(TTNN::Ops::Eltwise::Complex ALIAS ttnn_op_eltwise_complex)
 
-target_precompile_headers(ttnn_op_eltwise_complex REUSE_FROM TT::CommonPCH)
+if(TT_METAL_BUILD_AGGRESSIVE_PCH)
+    target_precompile_headers(ttnn_op_eltwise_complex REUSE_FROM ttnn_core)
+else()
+    target_precompile_headers(ttnn_op_eltwise_complex REUSE_FROM TT::CommonPCH)
+endif()
 TT_ENABLE_UNITY_BUILD(ttnn_op_eltwise_complex)
 
 target_sources(