--- a/ttnn/pch.hpp
+++ b/ttnn/pch.hpp
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
 #pragma once
 
 #include "../std_pch.hpp"