--- a/ttnn/pybind_pch.hpp
+++ b/ttnn/pybind_pch.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#import <pybind11/cast.h>
+#import <pybind11/numpy.h>
+#import <pybind11/operators.h>
+#import <pybind11/pybind11.h>
+#import <pybind11/pytypes.h>
+#import <pybind11/stl.h>
+#import <pybind11/stl_bind.h>