… ttnn/cpp/ttnn/operations/eltwise/quantization/CMakeLists.txt
… ttnn/cpp/ttnn/operations/eltwise/ternary/CMakeLists.txt
… ttnn/cpp/ttnn/operations/eltwise/ternary_backward/CMakeLists.txt
… ttnn/cpp/ttnn/operations/eltwise/unary/CMakeLists.txt
… ttnn/cpp/ttnn/operations/eltwise/unary_backward/CMakeLists.txt
… ttnn/cpp/ttnn/operations/embedding/CMakeLists.txt
… ttnn/cpp/ttnn/operations/embedding_backward/CMakeLists.txt
… ttnn/cpp/ttnn/operations/examples/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/adaptive_pool/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/auto_format/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/bcast_to/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/ccl/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/cnn/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/conv3d/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/copy/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/dropout/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/matmul/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/paged_cache/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/plusone/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/reduction/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/reshape/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/ssm/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/test/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/transformer/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/unary_backward/CMakeLists.txt
… ttnn/cpp/ttnn/operations/experimental/where/CMakeLists.txt
… ttnn/cpp/ttnn/operations/full/CMakeLists.txt
… ttnn/cpp/ttnn/operations/full_like/CMakeLists.txt
… ttnn/cpp/ttnn/operations/index_fill/CMakeLists.txt
… ttnn/cpp/ttnn/operations/kv_cache/CMakeLists.txt