Currently in repo, there is a `TT::CommonPCH` that makes precompile header for many projects.
This PR will add more header to be included in that PCH file that improves build time by 10% if you build with default arguments.

On the other hand, for even more increase in build time, one can use `--build-aggressive-pch` argument to `build_metal.sh` which will cause multiple PCH files to be created and changing include files of some projects will trigger full rebuild of other projects (instead of just many files to be rebuilt). From what I've tested, `--build-aggressive-pch` reduces build times by `40%` if you build with default arguments. I assume that safe bet would be using `--build-aggressive-pch` in CI (or that might clash with ccache).

Files with slowest build times:
```
39.125s   ttnn/CMakeFiles/ttnn.dir/cpp/ttnn/operations/eltwise/binary/binary_pybind.cpp.o
35.693s   ttnn/CMakeFiles/ttnn.dir/cpp/ttnn/operations/eltwise/unary/unary_pybind.cpp.o
26.039s   ttnn/CMakeFiles/ttnn.dir/cpp/ttnn/operations/eltwise/unary_backward/unary_backward_pybind.cpp.o
15.426s   ttnn/CMakeFiles/ttnn.dir/cpp/ttnn/operations/matmul/matmul_pybind.cpp.o
14.558s   ttnn/CMakeFiles/ttnn.dir/cpp/ttnn/operations/eltwise/binary_backward/binary_backward_pybind.cpp.o
14.480s   ttnn/CMakeFiles/ttnn.dir/cpp/ttnn-pybind/tensor.cpp.o
14.387s   ttnn/CMakeFiles/ttnn_core.dir/core/tensor/xtensor/partition.cpp.o
13.992s   ttnn/CMakeFiles/ttnn_core.dir/core/graph/graph_argument_serializer.cpp.o
```